+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 ### Added
 
 - `Reverse` is added to the string builder
+- `Substring` is added to the string builder
 
 ## [0.9.0] - 2023-08-26